requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
pypdf>=3.15.0 